[264. Ugly Number II][Compile Error]committed by LeetCode Extension

+++ 264-Ugly-Number-II/solution.java
@@ -1,9 +1,9 @@
 public class Solution {
     public int nthUglyNumber(int n) {
         if(n==1) return 1;
-        PriorityQueue<Long> q = new PriorityQueue<Long>();
+        PriorityQueue<Integer> q = new PriorityQueue<Integer>();
         q.add(1l);
-        long temp=0;
+        int temp=0;
         for(int i=1; i<n; i++)
         {
             temp = q.poll();
